pyproject.toml
[project]
name = "arpes"
version = "4.0.0-beta0"
description = "PyARPES: An analysis framework for multimodal angle-resolved photoemission spectroscopies"
authors = [
  { name = "Conrad Stansbury", email = "chstan@berkeley.edu" },
  { name = "Ryuichi Arafune", email = "ryuichi.arafune@gmail.com" },
]
readme = "README.rst"
requires-python = ">= 3.10"

dependencies = [
  "holoviews",
  "astropy>=5.3",
  "xarray>=2023.5.0",
  "lmfit>=1.2.1",
  "matplotlib>=3.8.0",
  "pint>=0.22",
  "h5py>=3.2.1",
  "ase>=3.22.1",
  "PySide6>=6.2.2",
  "dill>=0.3.6",
  "tqdm>=4.65.0",
  "numba>=0.57.1",
  "pyqtgraph>=0.12.4",
  "igorpy @ git+https://github.com/arafune/igorpy",
  "scikit-learn>=1.3.1",
  "scikit-image>=0.20.0",
  "netcdf4>=1.6.4",
  "rx>=3.2.0",
  "ipywidgets",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.rye]
managed = true
dev-dependencies = [
  "colorama>=0.4.6",
  "black>=23.3.0",
  "pytest>=7.3.2",
  "pytest-cov",
  "pytest-qt>=4.2.0",
  "types-tqdm>=4.65.0.1",
  "pluggy",
  "sphinx",
  "sphinxcontrib-restbuilder",
  "sphinx_rtd_theme",
  "nbsphinx",
  "sphinx_copybutton",
  "sphinxnotes-strike",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/arpes"]

[tool.coverge.run]
omit = [
  "tests/*",
  "src/arpes/plotting/*",
  "src/arpes/deep_learning/*",
  "src/arpes/widgets.py",
]

[tool.coverge.report]
exclude_lines = ["raise NotImplementedError", "pragma: no cover"]

[tool.pyright]
typeCheckingMode = "off"

pythonVersion = "3.11"
pythonPlatform = "All"

[tool.ruff]
lint.ignore = [
  "PD",     # pandas-vet
  "ANN101", # missing-type-self (ANN101)
  "N802",   # invalid-function-name (N802)
  "N806",   # non-lowercase-variable-in-function 
  "N999",   # invalid-module-name (N999)
  "S101",   # assert (S101)
  "TD002",  # missing-todo-author 
  "TD003",  # missing-todo-link
  "PD011",  # pandas-use-of-dot-values
  "FBT002", # boolean-default-value-in-function-definition
  "FIX002", # line-contains-todo
  "G004",   # logging-f-string
  #
  "NPY201", #  Numpy 2.0, 
]
target-version = "py310"
lint.select = ["ALL"]
line-length = 100

exclude = ["scripts", "docs", "conda"]

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]                         # unused-import
"src/arpes/__init__.py" = ["T201"]               # print used
"src/arpes/experiment/__init__.py" = ["ALL"]
"src/arpes/workflow.py" = ["T201", "T203"]
"src/arpes/fits/lmfit_html_repr.py" = ["SLF001"] # private member accecced

[tool.mypy]
ignore_missing_imports = true

[tool.black]
line-length = 100

[tool.commitizen]
name = "cz_emoji"
version = "0.0.1"
tag_format = "$version"

[tool.coverage.report]
exclude_also = ["if TYPE_CHECKING:"]
